darray: change element_size to size_t
This argument is usually the result of a sizeof().

+++ libnopegl/src/darray.h
@@ -23,17 +23,18 @@
 #define DARRAY_H
 
 #include <stdint.h>
+#include <stdlib.h>
 
 struct darray {
     uint8_t *data;
     int count;
     int capacity;
-    int element_size;
+    size_t element_size;
     int (*reserve)(struct darray *darray, int capacity);
     void (*release)(void *ptr);
 };
 
-void ngli_darray_init(struct darray *darray, int element_size, int aligned);
+void ngli_darray_init(struct darray *darray, size_t element_size, int aligned);
 void *ngli_darray_push(struct darray *darray, const void *element);
 void *ngli_darray_pop(struct darray *darray);
 void *ngli_darray_pop_unsafe(struct darray *darray);
